+++ packages/deepsea-tools/src/utils/postitiveIntParser.ts
@@ -1,0 +1,6 @@
+import { intParser } from "./intParser"
+
+export function postitiveIntParser(value: string | null | undefined) {
+    const num = intParser(value)
+    return num && num > 0 ? num : undefined
+}