+++ packages/soda-hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-hooks
 
+## 5.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - deepsea-tools@5.3.0
+
 ## 5.1.4
 
 ### Patch Changes